refactor(RenameProvider): вынести сборку WorkspaceEdit в отдельный класс
Сборка documentChanges/changes-map и аннотирование правок через ChangeAnnotation
перенесены в новый @Component RenameWorkspaceEditBuilder в пакете providers.
RenameProvider остаётся с резолвом символа, валидацией newName, guard'ом модуля и
сбором Map<uri, List<TextEdit>>, делегируя построение WorkspaceEdit билдеру с
передачей клиентских флагов documentChanges/changeAnnotationSupport параметрами.
Ресурс renameAnnotation перенесён в бандл нового класса. Поведение не изменилось,
число зависимостей RenameProvider опущено ниже порога S1200.
