+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/providers/RenameProvider.java
@@ -33,24 +33,17 @@ import com.github._1c_syntax.bsl.languageserver.utils.Resources;
 import com.github._1c_syntax.bsl.parser.BSLLexer;
 import com.github._1c_syntax.bsl.parser.BSLTokenizer;
 import lombok.RequiredArgsConstructor;
-import org.eclipse.lsp4j.AnnotatedTextEdit;
-import org.eclipse.lsp4j.ChangeAnnotation;
 import org.eclipse.lsp4j.ClientCapabilities;
 import org.eclipse.lsp4j.Location;
 import org.eclipse.lsp4j.Range;
 import org.eclipse.lsp4j.RenameParams;
-import org.eclipse.lsp4j.ResourceOperation;
-import org.eclipse.lsp4j.SnippetTextEdit;
 import org.eclipse.lsp4j.SymbolKind;
-import org.eclipse.lsp4j.TextDocumentEdit;
 import org.eclipse.lsp4j.TextDocumentPositionParams;
 import org.eclipse.lsp4j.TextEdit;
-import org.eclipse.lsp4j.VersionedTextDocumentIdentifier;
 import org.eclipse.lsp4j.WorkspaceClientCapabilities;
 import org.eclipse.lsp4j.WorkspaceEdit;
 import org.eclipse.lsp4j.WorkspaceEditCapabilities;
 import org.eclipse.lsp4j.jsonrpc.ResponseErrorException;
-import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.eclipse.lsp4j.jsonrpc.messages.ResponseError;
 import org.eclipse.lsp4j.jsonrpc.messages.ResponseErrorCode;
 import org.jspecify.annotations.Nullable;
@@ -60,7 +53,6 @@ import org.springframework.stereotype.Component;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 import java.util.stream.Collector;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -80,6 +72,7 @@ public final class RenameProvider {
   private final ReferenceIndex referenceIndex;
   private final Resources resources;
   private final ClientCapabilitiesHolder clientCapabilitiesHolder;
+  private final RenameWorkspaceEditBuilder workspaceEditBuilder;
 
   // Кэшируются на initialize. documentChanges — gate для построения WorkspaceEdit на
   // documentChanges (List<TextDocumentEdit>) вместо legacy changes-map; changeAnnotationSupport —
@@ -112,12 +105,11 @@ public final class RenameProvider {
   /**
    * Построить {@link WorkspaceEdit} с правками переименования символа.
    * <p>
-   * Если клиент заявил {@code workspace.workspaceEdit.documentChanges}, правки группируются по
-   * документам в {@code documentChanges} ({@link TextDocumentEdit} с версионированным
-   * идентификатором документа); при дополнительной поддержке
-   * {@code workspace.workspaceEdit.changeAnnotationSupport} правки оборачиваются в
-   * {@link AnnotatedTextEdit} и связываются с {@link ChangeAnnotation}, описывающей переименование.
-   * Иначе результат понижается до legacy {@code changes}-map для обратной совместимости.
+   * Резолвит переименовываемый символ, собирает текстовые правки по всем его вхождениям и
+   * делегирует выбор формата результата в {@link RenameWorkspaceEditBuilder}: при поддержке
+   * клиентом {@code workspace.workspaceEdit.documentChanges} (и опционально
+   * {@code workspace.workspaceEdit.changeAnnotationSupport}) возвращаются {@code documentChanges}
+   * с аннотациями правок, иначе результат понижается до legacy {@code changes}-map.
    *
    * @param documentContext Контекст документа.
    * @param params          Параметры вызова.
@@ -141,50 +133,14 @@ public final class RenameProvider {
         .stream().map(RenameProvider::referenceOf)
     ).collect(Collectors.groupingBy(ref -> ref.uri().toString(), getTexEdits(params)));
 
-    if (!documentChangesSupport) {
-      return new WorkspaceEdit(changes);
-    }
-
     var oldName = sourceDefinedSymbol.map(SourceDefinedSymbol::getName).orElse(params.getNewName());
-    return buildDocumentChanges(changes, oldName, params.getNewName());
-  }
-
-  private WorkspaceEdit buildDocumentChanges(
-    Map<String, List<TextEdit>> changes,
-    String oldName,
-    String newName
-  ) {
-    var workspaceEdit = new WorkspaceEdit();
-
-    var annotationId = "";
-    if (changeAnnotationSupport && !changes.isEmpty()) {
-      annotationId = UUID.randomUUID().toString();
-      var label = resources.getResourceString(getClass(), "renameAnnotation", oldName, newName);
-      workspaceEdit.setChangeAnnotations(Map.of(annotationId, new ChangeAnnotation(label)));
-    }
-
-    var resolvedAnnotationId = annotationId;
-    List<Either<TextDocumentEdit, ResourceOperation>> documentChanges = changes.entrySet().stream()
-      .map((Map.Entry<String, List<TextEdit>> entry) -> {
-        var textDocument = new VersionedTextDocumentIdentifier(entry.getKey(), null);
-        var edits = annotate(entry.getValue(), resolvedAnnotationId);
-        return Either.<TextDocumentEdit, ResourceOperation>forLeft(new TextDocumentEdit(textDocument, edits));
-      })
-      .toList();
-
-    workspaceEdit.setDocumentChanges(documentChanges);
-    return workspaceEdit;
-  }
-
-  private static List<Either<TextEdit, SnippetTextEdit>> annotate(List<TextEdit> textEdits, String annotationId) {
-    return textEdits.stream()
-      .map((TextEdit textEdit) -> {
-        TextEdit edit = annotationId.isEmpty()
-          ? textEdit
-          : new AnnotatedTextEdit(textEdit.getRange(), textEdit.getNewText(), annotationId);
-        return Either.<TextEdit, SnippetTextEdit>forLeft(edit);
-      })
-      .toList();
+    return workspaceEditBuilder.build(
+      changes,
+      oldName,
+      params.getNewName(),
+      documentChangesSupport,
+      changeAnnotationSupport
+    );
   }
 
   private static Reference referenceOf(SourceDefinedSymbol symbol) {
