+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/providers/RenameProvider_en.properties
@@ -1,2 +1,1 @@
 invalidNewName=Invalid name for rename: "%s". The name must be a valid identifier and must not be a keyword.
-renameAnnotation=Rename: %s -> %s
