+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/providers/RenameProvider_ru.properties
@@ -1,2 +1,1 @@
 invalidNewName=Некорректное имя для переименования: "%s". Имя должно быть корректным идентификатором и не совпадать с ключевым словом.
-renameAnnotation=Переименование: %s -> %s
