+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/providers/RenameWorkspaceEditBuilder_en.properties
@@ -1,0 +1,1 @@
+renameAnnotation=Rename: %s -> %s
