+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/providers/RenameWorkspaceEditBuilder_ru.properties
@@ -1,0 +1,1 @@
+renameAnnotation=Переименование: %s -> %s
